backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-socketio[asgi]==5.11.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
alembic==1.13.3
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.17
email-validator==2.2.0
cryptography==43.0.3
PyJWT==2.10.1
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.25.2
anyio==4.6.2.post1
python-dateutil==2.9.0.post0